Moved modals to their own folder
